Make backgorund process for notifications

+++ app/src/main/java/com/example/myapplication/Data/Expense.java
@@ -7,12 +7,12 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Expense {
 
-    public Expense()
-    {}
-    public Expense(String type, int amount ,long currentDate , String subType)
-    {
-        this.type=type;
-        this.amount=amount;
+    public Expense() {
+    }
+
+    public Expense(String type, float amount, long currentDate, String subType) {
+        this.type = type;
+        this.amount = amount;
         this.currentDate = currentDate;
         this.subType = subType;
     }
@@ -20,24 +20,23 @@ public class Expense {
     @PrimaryKey(autoGenerate = true)
     public int exid;
 
-    @ColumnInfo(name="type")
+    @ColumnInfo(name = "type")
     public String type;
 
-    @ColumnInfo(name="subType")
+    @ColumnInfo(name = "subType")
     public String subType;
 
 
-    @ColumnInfo(name="amount")
-    public int amount;
+    @ColumnInfo(name = "amount")
+    public float amount;
 
-    @ColumnInfo(name="currentDate")
+    @ColumnInfo(name = "currentDate")
     public long currentDate;
 
-    @ColumnInfo(name="uid")
+    @ColumnInfo(name = "uid")
     public int uid;
 
-    public long getCurrentDate()
-    {
+    public long getCurrentDate() {
         return this.currentDate;
     }
